reduce floats until we move to avalonia vectors

+++ Avalonia.Lottie/AccelerateDecelerateInterpolator.cs
@@ -4,13 +4,13 @@ namespace Avalonia.Lottie
 {
     internal class AccelerateDecelerateInterpolator : IInterpolator
     {
-        public float GetInterpolation(float f)
+        public double  GetInterpolation(double f)
         {
-            if (f < 0 || float.IsNaN(f))
+            if (f < 0 || double .IsNaN(f))
                 f = 0;
             if (f > 1)
                 f = 1;
-            return (float) (Math.Cos((f + 1) * Math.PI) / 2 + 0.5);
+            return  (Math.Cos((f + 1) * Math.PI) / 2 + 0.5);
         }
     }
 }
